Add test to detect some rendering inaccuracy.
When the decoded frame has a smaller size than the buffer, the decoder
will send cropping information to SurfaceTexture. In this case, the
output from getTransformMatrix will cause rendering to be inaccurate
(due
to the workaround to prevent texture bleeding). DecodeAccuracyTest was
not aware of this inaccuracy and would fail even when the decoder
produced the correct output buffer. This change makes DecodeAccuracyTest
more lenient by testing the rendered output against 6 cropped and scaled
versions of the "golden" image in addition to the original. The six ways
of cropping and scaling are based on GLConsumer.computeTransformMatrix.

Bug: 29218487

Change-Id: Iac42a0402c0086ee7a50649a7dbcb7ffd9de2ca2

Former-commit-id: d030230a35b54e0017cae83e5312212148b328e9