Clean up CTS MediaPreparer
Prevent case in which media folder is created but not populated (due to
interruption in a previous run) and results in a TargetSetupError in the
next run. Create the temporary zip file in the tmp directory, instead of
the current directory. Also make the code more readable and provide
better logging.

Bug:30120206
Change-Id: Ia5b5545aeb88ccd627f84aa605c5bb03acc47da2

Former-commit-id: 69c9ee98999f3ecf12590605475bb666cbef283b