Restore font scale to default in Settings_SystemTest
In case an assertion fails before restoring the default font scale
Settings_SystemTest may leave the system in a state where font scale is
very big, which breaks some assumptions in text related tests. This CL
is an attempt to reduce the likelihood of test flakiness. It tries to
restore the font scale in a finally block, and also reduces the amount
of font scale that is tested from 10.0f to 1.2f.

Bug: 29569812
Change-Id: I9d7113c89f3b5ee6903980fe4eeca556192e0018

Former-commit-id: 3835bb7137fd31f99d256f75163d80a829bd40d5