Third pass for completing CTS coverage of RemoteViews
* The changes in RemoteViewsTest cover the simpler cases that operate
on "regular" views.
* The new RemoteViewsWidgetTest is for APIs that can only operate on
content that is hosted in AppWidgetViewHost. Specifically, the inner
SetRemoteViewsAdapterList does class check for the root host. Since
the only remotable widget that allows testing for setDisplayedChild,
showNext and showPrevious is StackView, and the only way to configure
remote content of StackView is through setRemoteAdapter, we have to
test these APIs in a widget environment.
* The overall flow of configuring the appwidget environment is lifted
from AppWidgetTest CTS class, with minor simplifications along the way.
In addition, the usage of libcore.io APIs is replaced with more stable
alternatives.

Bug: 29805243
Change-Id: I402c468a6c5dd12eb7a03b78871898694653b448

Former-commit-id: f45d6ea05c42603863cff45ab339292084e99f45