Remove WebView startup StrictMode CTS test and add NullWebView check
The WebView startup StrictMode test fails for Wear, and passes
incorrectly when run as part of a series of tests anyway so there is no
reason to keep it as is (if implemented correctly, or just run on its
own, it will fail for a non-wear device).

Also add NullWebView check to a WebView ServiceWorker test to early-out
if WebView is not available on the device.

Bug: 30119659
Change-Id: I2a60d02175c5085704285e570a93fe4dc00b57db

Former-commit-id: 596d54378d303231a330fcbbc8bc247e6955d310