Fix android.telecom.cts.ConferenceTest#testConnectionProperties
This test was failing and taking the rest of the tests along with it
because PROPERTY_IS_EXTERNAL_CALL caused Telecom's InCallController to
disregard the conference call, which led to the disconnect request being
ignored and subsequently the call being orphaned in Telecom.
Changed to an unrelated property since the test did not actually test
any behavior specific to the property.

Change-Id: I986f177f37ca23da946add4501074f0a3205989f
Fix: 30087065

Former-commit-id: d763a665e82c449868ccab1c0f92318fffd03036