Fix CTS Theme tests
These tests depended on a folder with the expected screenshots
according to API level, and we only had up to 24

Given that CTS tests only run for the current sdk level, this
change removes the different API level folders. The new golden
images are the same as were in the 24 folder, no changes have
been made for API 25.

Bug: 30216089
Change-Id: If653212acddf1324fcc447d434033cf623720bd0

Former-commit-id: eeed0867bebe6ffb57a37f330700c7198bb5cc6d