Add a host-side test to verify /proc/stat exists.
Android Studio Performance Tools needs, as the shell user, to read
/proc/stat file to collect live metrics data. It requires this
file is in a reasonable shape, and being actively updated.

I have verified this test passing on a Nexus 6P running O. I have
also verified this test passing on a non-Nexus device (Samsung 6
running 6.0.1). In order to run the test on the non-Nexus device,
I have ported this CL into mnc-dev branch (http://ag/1282570) which
will not be submitted. I will only submit this CL on master branch.

http://b/30595738 is tracking this work.

Change-Id: I4ba82662b2e4b263fb732082238940f9e135ce48

Former-commit-id: 1cb36ce3333678fa5700abfa10cffd5b64d199c9