CP: Screen on and stay on for GLES tests
NOTE: Required a change into the API level in the build file to allow
the 64bit build to happen.

Bug: 34731040
Test: Build & run CtsOpenGLTestCases on N5X with the screen off.

Original commit:

272d01f05bd6846eaa79acb4322801aaeb5101fe [formerly ca1898d9472b3d252de2678ed925d872edf1494e]

Screen on and stay on for GLES tests

Add window flags that ensure that keyguard is opened, the display is
turned on, and stays on for the duration of the test.

Also, fix the target SDK to 14 mirroring manifest.xml minSdk

Bug: 29381895
Change-Id: Ia96828b436b962d9850a68583136c98bbc5dfbf2

Former-commit-id: 533642ba606a228851b7bf22df652f16432c9b14