Fix MultiFloat tests
Current tests for multi-float animators assume too much about how the
intermediate values are calculated. Floating-point roundoff errors make
the comparisons to internal and test calculations ripe for failure.
The fix is to simply use an epsilong to account for this error (as long as
the calculation is close enough, that's sufficient for these tests).

Issue #30769814 ofFloatPath CTS tests broken

Change-Id: Ifbe5f855a05d799d15f96f4fea679d2af4e079e9

Former-commit-id: 64fe275fe1127cc8c0353cbe97b9156db7897a5f