DO NOT MERGE: Improve task list parsing in CrossAppDragAndDropTests
In order to find a specific activity CrossAppDragAndDropTests looks
for this activity's name in the output of "am stack list" command.

Turns out that in addition to the usual place (the second token
in the line) an activity name may also occur in lines related
to other activities (after the "topActivity" key). This breaks the test.

This is fixed by looking for the activity name only before the first
occurrence of "topActivity" substring.

Bug: 30919524
Change-Id: I112f0ea3997177032882984b97dcf9212f327cec

Former-commit-id: 23f7892190d294e2152fbc1d9571ea38a1a35228