Disable transition animation for AVDParemeterizedTests
This CL disables the transition animation by setting the scale to
0 before the test class and restoring the scale after all the tests
in AVDParameterizedTests class. This fixes the issue where consecutive
screenshots are taken during transition animation and subsequently
fail the tests as there's no change in the screenshots.

This CL also changes the start value of the AVD fill color such that
all 3 color channels are animating, producing a more prominant delta
in consecutive screenshots.

BUG: 31061872
Test: This CL.

Change-Id: I7e90e1cf9b7e12742e382a0ad81bf444fc603a01

Former-commit-id: 60e4f2500b0304a4351d353018a7633d1fb4e8b7