Fix subplan filtering
For filtering based on previous results, explicitly exclude modules that
contain 0 tests that should be rerun. This is especially necessary when
not-executed results should not be rerun, since these tests are not in
the result and cannot be excluded by filters directly. This also
prevents the loading of modules that contain no tests post-filtering.

Also rename -r option for subplan creation to conflict less with the
previous session's 'result', and modify the filter collections within
the compatibility testing infrastructure to be sets, rather than lists.

bug:29736541
Change-Id: I425c502c2565cdbe597da52321f5144dc532290f

Former-commit-id: d3a3d7c3aae384c383260ff11e6eabd981d8698b