Test: fix WebView testFindNext flakyness
testFindNext is currently flaky (failing 10-20% of time). The issue is findAll
is called before page is fully loaded and FindResultsListener callback
with number of matches being 0. This fix is to allow FindResultsListener to call
findAll again is the match number is not ideal. more detail of the bug at
crbug.com/618089

Change-Id: I8f647af7e2b566da123165f842e0a7f6ed79d482

Former-commit-id: ea1e7f02785e1f4ab22e68a8745fe5cbf633a8ea