Fix testTimeTrackingAPI_ChainedActivityExit test
Test fails when ran alone (at least on emulators). It expects broadcast
ACTIVITY_LAUNCHED_ACTION, which is never sent from the two activities
that test launches. Change the broadcast filter to listen to the
ACTIVITY_CHAIN_EXIT_ACTION action, which was probably intended in the
first place (defined in the file, but never used).

Change-Id: Ib6d267ca7161131ae5c8048b12bbd06304538795

Former-commit-id: 35f0ffed42132e7a5d9341f0606e452f25ea20e0