dram: Fix ArrayIndexOutOfBoundsException when all data are rejected.
BandwidthTest will reject data that differs from median by 10%, but in
some devices, bandwidth may vary so much (10x) that all data are rejected
by getStatWithOutlierRejection which results in an exception.
This CL makes the data length odd so that the median is just the number
in the middle of the sorted data instead of an average of two numbers in
the middle, this way the median will never be rejected.

Bug: 31606924
Test: cts-tradefed run cts --module CtsDramTestCases --test android.dram.cts.BandwidthTest

Change-Id: I677e4cb5b438eedef33e99081aea97e41e55f925

Former-commit-id: 47fb62ade03084e212e9f7a297c55e058dda69bc