Magnification center moved to the screen center
In
android.accessibilityservice.cts.AccessibilityGestureDispatchTest#testClickWhenMagnified_matchesActualTouch
magnification center was moved to the screen center. Previously
magnification was align so that top left corner was mapped to itself.
It worked fine on rectangular screens, but was failing on round screens
as top left point with its surroundings was off screen.

Bug: 31985427
Change-Id: I8d9d6d0ac4797ffa7c4c098814dcec518a1a77f6

Former-commit-id: c49f3f13d74d2d2d4ae2a00808f4f008663d6729