CTS test for SecurityLog.SecurityEvent
Test is not too thorough since we cannot assume anything about those
events. A more comprehensive test will be done as a part of AfW test
harness.

The test is a bit time consuming: it reboots the device and waits for 70
seconds to get logs. (this interval is an implementation detail, but
there is no better way currently.)

Fixes:30495467
Test: cts-tradefed run cts -m CtsDevicePolicyManagerTestCases -t
com.android.cts.devicepolicy.DeviceOwnerTest#testSecurityLoggingWithSingleUser
Change-Id: I535c2f1e6deb72187bc785eb3092dc8341c84888

Former-commit-id: bb6f8c5bb585c4df1f036e3855636e2bfda34245