DO NOT MERGE SurfaceViewSyncTests: survive permission activity recreation.
Following race happens: permission activity appears on a screen,
uiautomation finds accept button on a screen and we successfully click it,
however, in parallel system decides to recreate the permission activity due to configChange,
and, actually, our click on the button doesn't grant us a permission.

Solution is ugly hack: click that button until we actually get a permission.

bug: 32244764
Change-Id: Iabe0ab5a2854464ccf314e0da16accdb6a084efe

Former-commit-id: ef5f2bcb732550f178cb97b266b85acda3f53a2b