Fix CtsAadbHostTestCases#testSyncFiles_extStorageVariable
A device has its own clock and timezone which can be off from the
host on which we run the test.

A file could get synchronized twice if the lastModified timestamp
was lower or within a certain leeway of the lastModified time on the
device.

This test will fail if the second synchronization results in a timestamp
with a different minute. An example is where the first sync happens at
11:00:59, and the next sync happens at 11:01:00

Bug: 32409804
Change-Id: Ie8d26ea536cb29ec6316cb0bf05ce82f47074a15

Former-commit-id: 9e2cbdda3228e47c5f33d09171628e9b21460b08