CTS for MenuItem shortcut modifiers
MenuItem now allows specifying modifiers for keyboard shortcuts. To test
this functionality, the commit

1. Adds testInflateShortcutModifiersFromXml to MenuInflaterTest
This test makes sure that modifiers are picked up and set up correctly
when a menu xml in inflated.

2. Adds testPerformShortcut to newly created MenuTest
This test makes sure that the correct shortcut is in fact triggered when
the correponding KeyEvent is sent.

Bug: 32165848
Test: Run the tests using following commands:
> cts-tradefed run cts-dev -m CtsViewTestCases -t android.view.cts.MenuTest#testPerformShortcut
> cts-tradefed run cts-dev -m CtsViewTestCases -t android.view.cts.MenuInflaterTest#testInflateShortcutModifiersFromXml
Change-Id: I9d261636a9b68292eaed403085373a13b5679df3

Former-commit-id: e2221b10490bde5395e3bd456161c807af166290