Made TestFilter usable with AndroidJUnitRunner
Renamed TestFilter to something more meaningful (all Filter
implementations filter tests) and made it work standalone with
AndroidJUnitRunner by changing the constructor to take a
Bundle and constructing the expectation store in there rather
than pass it in from CoreTestRunner.

Made sure when testing that the number of tests that are run
was the same as the number run before making this change.

Bug: 28689581
Test: Built cts, ran cts-tradefed run cts -m CtsLibcoreTestCases
Change-Id: I4feed6e9912b63f909aaf666aa8ff55b3e4e795c

Former-commit-id: a483016f9acdb4ac5d348e1fcdd177920dde4596