Include device side stacktrace in the hostside stacktrace
Key changes:
1. BaseDevicePolicyTest.runDeviceTests* now throw AssertionError
instead of return false for fail tests.

2. Device side stack trace is put in the detail error message in the
AssertionError

3. Move the failed reason from host side to device side. If the reason
is not meaningful, simply drop it.

Test: Try to write a fail test and I can observe the device side
stacktrace is embeded in the host side stacktrace
In theory, I should run every test to make sure it doesn't break anything.
But it will takes my the whole day to run all the CTS... I will monitor
the status of CTS after merging it.

Fix: 32468481

Change-Id: I7827a0568dc5675e02f0d6a27e51f8129081ca83

Former-commit-id: 229afe3d739e3aa68893134024a860705375870b