Devices may not support Direct Boot emulation.
Direct Boot emulation is encouraged to help developers test their
apps, but it's not required.  If a device responds with "Emulation
not supported" then we pass the test.

However, note that both testDirectBootNative() and
testDirectBootNone() absolutely must always pass, since they verify
the device behavior in whatever the factory default state is.

Test: cts-tradefed run cts-dev --module CtsAppSecurityHostTestCases --test android.appsecurity.cts.DirectBootHostTest#testDirectBootEmulated
Bug: 30686636
Change-Id: Ie18e42eda43a0f69c3e4f9e80b280014f90579ca

Former-commit-id: a596c33172f1631ca987f771e7a2d77114839fc3