Reduce feedEncoder loading when witholding buffer release
testReleaseAfterFlush aggressively loads the encoder with frames anytime
an output buffer becomes available.

On devices such as the emulator where the codec has a low buffer count
and slower performance, the test exhausts the buffers but continues to try
to feed the encoder which causes the test to hang.

To reduce this loading, the test will now only feed the encoder until
the first output buffer is recieved. It will then continue to accumulate
any output buffers generated.

A side effect will be that there will be fewer total number of output
buffers accumulated during the test.

Bug: 30843570
Change-Id: Ib724b2338b03501e1df3f6915820d514f3aa41bc

Former-commit-id: 23d745888de93d8cdb95883d100ff0be7e4121e8