Switched luni, okhttp, jsr166 and icu to use AndroidJUnitRunner
Ensured that the same tests were run before and after the
change. One slight change is that okhttp and jsr166 are no
longer running android.test.LoaderTestCase which is ok
because they should not be running that test anyway.

The cts-runner-specific-failures.txt file was removed because a
bug in AndroidJUnitRunner was fixed. The bug meant that while
scanning the APK, classes that implement junit.framework.Test,
but do not extend junit.framework.TestCase would be added to the
list of classes to run but would cause errors as they were
rejected by the RunnerBuilder implementations.

Bug: 28689581
Test: ran CtsLibcoreTestCases CtsLibcoreJsr166TestCases and CtsLibcoreOkHttpTestCases
Change-Id: Idc5cab2b3cfe93eb569e421e909e6909afba4d9c

Former-commit-id: aa3f3a307b2daa71bedff4160cf12a6b7a1cdc51