CTS: Sensor - Correct FrequencyVerificationTest Pass Range
Currently, FrequencyVerificationTest sets the pass range as 90 Hz to 440 Hz
and 0.9 Hz to 11 Hz for testSensor200Hz and testSensor1Hz respectively.
However, these ranges have not been set by fully taking care of minimum and
maximum rate a device supports. So for a device which supports 5 Hz as minimum
rate and 100 Hz as maximum rate then pass range for any requested sampling
period should not be outside of 4.5 Hz to 220 Hz limit. Therefore lets fix
the pass thresholds range.

And when sensor do not report its max delay or sampling rate is overloaded,
then we still keep lowerExpectedHz same as before but keep upperExpectedHz
to be not more than maxRate unlike before.

For example:
For a sensor with the HAL specification of 100Hz as its maximum rate,
TestSensor200Hz passes even if sensor generates data at a rate of 250 Hz
or even more, which ideally should fail as maximum possible allowed range is
actually under 220 Hz.

Test: Run testAccelerometer200Hz on a device which has specified 100Hz
as its max rate at HAL but hardware supports 200Hz as its max delay and
check if test passes or not.
So a failure would notify that either there is a need to change the max rate
in HAL specification or fix a bug in event generation.

Also, execute sensor cts tests and check if tests are pass as before.

Change-Id: I1b8574b5c7f79b8182c2cdfddd5d87e2043cb69d
Signed-off-by: Aniroop Mathur <a.mathur@samsung.com>

Former-commit-id: 97f197e5624829bfeacc0a0eb677200b8d1bccfe