Don't require WiFi in CtsHostsideNetworkTests
WiFi is not a CDD requirement so these tests should not fail when the
device under test does not have WiFi. The behavior is changed so that if
there is WiFi then both metered and unmetered tests will run. If there
is no WiFi and the current connection is metered then only metered tests
will run. If there is no WiFi and the current connection is not metered
then only unmetered tests will run.

Test: Successfully ran CTS test on both emulator and shamu.

BUG: 31648368
Change-Id: Ic643d2490e0a7e69b57a44599f1a4c57c67da873

Former-commit-id: 43fa4fc329d6ba6f1d405cb7e5c46166a2bded72