AutoSize TextView (part 6) - introduce autoSizeMaxTextSize [CTS]
autoSizeMaxTextSize attribute for TextView has been introduced
(instead of using getTextSize() as max size). Reflect this in
the test layout.

Also discovered flakiness in the existing tests and after
investigating it turns out that mokito does not play well
with multiple threads. This led to changing the tests from
verifying that setTextSize has been called to verifing the
text size has changed.

Bug: 33449596
Bug: 32221168
Test: cts-tradefed run cts-dev -m CtsWidgetTestCases -t \
      android.widget.cts.TextViewTest

Change-Id: Idb5224ccfc64b3d5e89cdfda6ee5834d3d96b584

Former-commit-id: 069b4a9d9beacb7ae994d97ef44b696917822d21