Use getLocationOnScreen to compute verification bitmap offset
The previous approach assumes that mViewWrapper is positioned at
(0,0), which is not true if the test is running in freeform mode
(with a window caption). Changing this to use getLocationOnScreen
allows the verification bitmap to be cropped from the screenshot
correctly.

Bug: 33414955
Test: run cts -m CtsUiRenderingTestCases. No additional failures
beyond existing failures. Tests now pass on manually forced
freeform mode.

Change-Id: I34843c7f7941ca70360e2ce114c7523ea5caf444

Former-commit-id: 0f9ceac0ac2a6b062af2b198f7350980a8c2b18d