Modify for cts test android.media.cts.EncodeVirtualDisplayWithCompositionTest#testVirtualDisplayRecycles
This commit is a suppliment of https://android-review.googlesource.com/#/c/194913/.
We just wanna keep the func getMaxSupportedEncoderSize() will use the same parameters
AS of the func doEncoding() when finding a encoder. E.g., When HW AVC 1080p is not
supported but there is one SW AVC Encoder in media_codec_google_video.xml with max
size 1920*1088 and bitrate range 1-12000000. getMaxSupportedSize() will return SW enc 1920*1080,
but doEncoding() won't return SW enc 1920*1080 whit bitrate value of 20000000 which exceeds
the max bitrate in MediaCodecList.


Former-commit-id: 0329617d7b104a8db48718a2638b07099736cde7