Add android.app.action.SET_NEW_PASSWORD support as prerequisite of DirectBootHostTest
DirectBootHostTest uses the android.app.action.SET_NEW_PASSWORD intent
to launch the password settings activity. Some devices provide full-disk
encryption and authentication outside of Android, and therefore can
never be run in a state where the device has been powered on but the
user has not unlocked the device.

Bug: 33754525
Test: run cts -m CtsAppSecurityHostTestCases -t android.appsecurity.cts.DirectBootHostTest
Change-Id: Ic37c74dabe6cfa30a25ac18618da06a09091f803

Former-commit-id: 995954fff0374ebf70a944a967d7f8b51874a894