Allow shared lib and package name to match - CTS
The old package manager implementation was guarding for
shared libs to not match any package name but due to a
bug this was not enfroced and as a result we have apps
where the package and lib match, e.g. GCore. The new
static shared lib support properly enforced the rule
breaking apps. So, this CL relexes the constraint.

Test: These are the tests

bug:34598313

Change-Id: I6883e0e05ff8d239de8ff7946a6725eca3fc04f6

Former-commit-id: 67c7138e7a74ea1556ea29b03d1abec4a141d08f