Replace strlen("foo") with sizeof("foo") - 1.
We have an upcoming change to Bionic that no longer allows us to treat
strlen("foo") as a constant expression, which causes this bit of code to
no longer compile. So, we need to either use __builtin_strlen("foo") or
sizeof("foo") - 1 instead. (Note that the *optimizer* can still turn it
into a constant, but optimization happens after we figure out if
something is actually a constant expression or not.)

sizeof("foo") is used elsewhere in this file, so I just went with that.

Bug: 32073964
Test: Now builds.
Change-Id: I2797ee75fd114e237de8e97c50549763c88f73f4

Former-commit-id: f14a44323f0f1ec62006ca80ee2e119290a5d7b1