Add instant cookie APIs - CTS
This change adds APIs for instant apps to store cookie data
that is presisted across instant installs and acorss the
upgrade from an instant to a standard app. Standard apps
can use the cookie APIs but when they are uninstalled the
cookie is also deleted. The cookies are kept longer than
the instant apps as they are much smaller - 16KB by default.
We can change the cookie size via a system setting i.e.
after we ship we can increase size if needed.

We also add internal APIs to surface information about
installed and uninstalled instant apps which should be
used for showing them in the UI. For this puporse we store
the icon, permissions, and label of unisntalled apps. If
the app is re-installed we drop this meta-data but keep
the cookie around.

Test: CTS tests pass; hiddent APIs tested manually

Change-Id: Iefcf1f0aae79756b7354789cd9b47c607aabda9f

Former-commit-id: 1e7292fe280d93b1eb74e3b6c3e4ec9dbb0f65e3