Add CTS tests for InCallService which can see self-managed calls.
Added new telecom3 cts package which includes an InCallService which
defines the meta-data indicating that it wants to see self-managed calls.

Added unbind latch to CtsConnectionService used to wait on unbinding of
the CTS CS.  This is to help improve reliability of unbind checking.

Moved a whole bunch of methods to TestUtils so they could be used in the
other telecom cts packages (used "refactor" in AS, hence the large number
of changes due to new class location).

Test: This is a test.
Bug: 34159263
Change-Id: I1a055acc097ec78d39913d246f396488f6f2f162

Former-commit-id: 1595c6a337805d838207d69a2f7aa2be27e5f21f