More CTS tests for AnimatorSet
This CL verifies AnimatorSet's reverse behavior when both
AnimatorSet and child animators have start delay set on
them. We want to make sure when revers()'ing, the start
delays are skipped.

Another 2 tests in this CL ensure 0 and non-zero duration
scales are correctly handled.

Finally, the new API getCurrentPlayTime() is tested in this
CL.

Test: This CL
BUG: 30993532

Change-Id: Ief519c4cac0409a08ad5d552093fbb5a7212dfb7

Former-commit-id: 72fda824bb1b503cbce5df410a35f48598a2b3a0