SurfaceViewSyncTests: survive permission activity recreation.
Following race happens: permission activity appears on a screen,
uiautomation finds accept button on a screen and we successfully click it,
however, in parallel system decides to recreate the permission activity due to configChange,
and, actually, our click on the button doesn't grant us a permission.

Solution is ugly hack: click that button until we actually get a permission.

Reapply fix from nyc-dev, which was lost because of merge conflicts.

bug: 34329728,  33702866
Change-Id: I1ab61e459ee87417ea08f494fa44be6e391e3a2f

Former-commit-id: 338980e981b0a2c7d61bc9d3b1952110b78a77b0