Fix VoicemailContractTest.testVoicemailsTable
When a row in the voicemail content table is modified by the owner
(SOURCE_PACKAGE) of that row, DIRTY should be set to 0.

Previously there are some bug in the provider preventing it from
identitfying the calling package and DIRTY was not overridden, so
the test is expecting the wrong value.

The value is corrected in this CL.

Change-Id: I6b0df4f89f2650c5ac8b44cfc0f60e2b2b0ffbc5
Fixes: 36195948
Test: CtsProviderTestCases VoicemailContractTest

Former-commit-id: 9a5662e7659c57cfa010dd46bd0068864e037754