Test color space conversions at draw time
Covers 8888, hardware and 565 cases, with opaque bitmaps, translucent
bitmaps and opaque bitmaps drawn with alpha. Each test is ran by drawing
the bitmap directly, by using it in a BitmapShader with a shape and by
using it in a BitmapShader with an ALPHA8 bitmap mask.

The tests also cover the various transfer function implementations
(hardware and parametric).

Test: CtsUiRenderingTests, CtsGraphicsTests
Bug: 32984164
Change-Id: I1202f7c6278a5e46de7bdc86373c7eb510d72df6

Former-commit-id: 71066bc2786b348a66cfcca5b1e6cad68c8c3450