CTS verifier: disk encrypted after provisioning
We add a test case in the managed provisioning CTS verifier test to
ensure that the device can be properly encrypted after the full
provisioning flow. The encryption status can by queried through
DevicePolicyManager#getStorageEncryptionStatus().

If the encryption status is ACTIVE or ACTIVE_PER_USER, the device is
considered encrypted and the test passes.  If the encryption status is
ACTIVE_DEFAULT_KEY, the tester will be asked to set a screen lock, after
which the test is run again.  Otherwise, the device is considered not
encrypted and the test fails.

The test is hidden for devices whose first API level is below N-MR1.

Test: make cts-verifier and run test
Bug: 27699209
Change-Id: Iea82c3affaa33952989fc9ca3743d521784c46a9

Former-commit-id: 1c3667a9dbd1070d535b58b169f8e7e0f7cc4e6f