Test correctness of TypedArray index list
The indices of TypedArray (gotten via TypedArray#getIndex(int))
should be asserted to contain exactly the expected set.

Previously, the API was being used incorrectly and a change in the
resource compiler broke the incorrect assumption that was being
asserted.

TypedArray#getIndex(int) takes a value in [0, TypedArray#getIndexCount())
and returns an offset into the R.styleable array that was used to create
the TypedArray.

The broken test was passing in the offset into R.styleable as the input,
which is incorrect but happened to work due to the way the resources
were ordered.

Bug: 36812266
Test: bit CtsContentTestCases:android.content.res.cts.TypedArrayTest
Change-Id: I183913bcbba9f668b8ebd048c8dd724a1733d578

Former-commit-id: 5c802b95e02b101ecf2cdbc0bb0a155c862b617c