Use dedicated cert for static shared lib tests
Packages using a static shared lib need to refer it by name,
version, and SHA-256 of signing cert. However, when CTS is
build for release the cert used to sign apps that do not
explicitly specify a signing cert changes as a result the
static shared lib tests fail as the app consumer refers to
the worng cert SHA-256. This change fixes the user cert.

Test: All static shared lib CTS tests now pass

bug:34712107

Change-Id: Ibfc37bf5f136d708bf449a24432e4f4c51f571b0

Former-commit-id: 40d4964104ae67c5e528854df9810ad27fb9ca42