Delete testAllCharacterDevicesAreSecure()
Delete the test testAllCharacterDevicesAreSecure(). Many of these tests
are already duplicated by SELinux neverallow rules, which are a cleaner
way to perform these kinds of tests and don't involve opening up stat()
permissions on various character device files. In addition, the normal
version of this test (not the host side version) doesn't work because
/dev isn't readable.

This test is a maintenance burden, doesn't meet our partner's needs, and
is a recurring source of waiver requests.

Bug: 36891574
Test: CTS compiles.
Change-Id: Ice9b9c32a87add3a0f16647a46af1be31ad5f044

Former-commit-id: 6243ff729f6652552812ac1198e06cba90fb2d67