Rework getMetrics() api per API council
Reworking the media metrics getMetrics() calls (currently in
MediaCodec, MediaExtractor, MediaPlayer, and MediaRecorder) to
fit new direction from the API Council.

Drop the MediaMetricsSet that we had in the first round; go back
to a PersistableBundle as the return type. Moves the key definitions
from MediaMetricsSet.MediaCodec to MediaCodec.MetricsConstants

The changes here in CTS track the changes to the API.

Bug: 37083862
Test: ran the corresponding CTS tests
Change-Id: I57908694f5219e214161e4e39d4efc77a5c6d78e

Former-commit-id: e11150941fb976252653648e617de89e96c948a1