Attempt to de-flaky GraphicStatsValidationTest
Fix argument ordering on assertEquals.

Switch to FrameMetricsObserver to listen
for real draw completed signal. This results
in spacing out frames on slower to devices
to avoid frame drops from queue stuffing
detection.

Strict assert number of frames on the activity
side to validate that the activity behaved appropriately
to determine if the flakiness is rendering-side or
reporting-side.

Test: this
Bug: 37438953

Change-Id: Ia3e8e51b1df92da15542915b2537ccce528d67bc

Former-commit-id: e8da3ac6297928d465933d13fb348c0a9bdd1921