Remove CTS tests for VR virtual displays temporarily
There is instability in VR virtual displays currently being experienced
in b/37519234. While this is being worked on, the interim workaround is
to comment out these CTS tests. The VR virtual display tests will be
re-enabled as a part of b/37576999.

Bug: 37519234
Test: > run cts -m CtsServicesHostTestCases
      - Passes all ActivityManagerDisplay test and prior tests.
      - Using ag/2098655 to avoid auto merge on master.

Merged-In: Iae7101bb1ab852097acb9e9f587249db4195fcfe
Change-Id: I0b7ecf8f0e0ca5cf64b87938c4aa71b5165de229

Former-commit-id: 41927d5dfb2dc349992889ec97c59d697a5231e4