Add test for resolving inline string for theme attribute
This was fixed in O in commit 7a37b74d37ff79e805c9e97d977e07bfec753c5a
but has caused some apps relying on the broken behaviour to crash.

Surprisingly, before the fix, an inline string value set for
an application attribute in the theme would not have the correct assetCookie
in the returned TypedArray. This would result in the string being
looked-up in the wrong StringBlock, yielding potential
IndexOutOfBoundsExceptions and incorrect values.

This adds a CTS test to ensure that the behaviour is correct for inline
strings.

Bug: 36379380
Test: bit CtsContentTestCases:android.content.res.cts.Resources_ThemeTest
Change-Id: I10a79a1d9a6d5008cdb92ee322aa0cb838f9e716

Former-commit-id: 43c96918333ebb564c4cf29937ad426c4a1c14a2