Remove view from parent before adding it as header/footer
ListView#addHeader/FooterView unintentionally accepts a view which is
other view's child. And it eventually causes ClassCastException. The
methods should throw IllegalStateException for the case. Before fixing
the issue in the framework, the CL stops adding other's child as a
header/footer so that CTS test will not be broken due to the framework
behavior change.

Bug: 37726743
Test: ListViewTest
Change-Id: I4ed93259b65bbf6a3ee87154e8716d8053b3c2f8

Former-commit-id: 6c96508ebb70c59de5ac841346a9f1e6c9dfd1cd