Move testWifiUpload next to testWifiDownload
The CPU frequency tests were recently added between testWifiDownload and
testWifiUpload. Logically, they should be adjacent. But this apparantly
isn't just an asthetic issue: currently, the CPU frequency tests fail,
and this seems to cause the testWifiUpload, which is called immediately
afterwards, to fail too (due to it not entering a background proc
state). Moving testWifiUpload will make it independent of the cpu
frequency tests failing.

Bug: 38397188
Test: run cts-dev -m CtsIncidentHostTestCases -t com.android.server.cts.BatteryStatsValidationTest
Change-Id: I32a18efcd7409237c3a15b58b9821ce617991f34

Former-commit-id: 47dd7cd325f32598c839e0e73c69834630fceb80