cts: fix potential crash due to race between close() and callbacks
MediaCodec callbacks, when set without a handler specified, will
run on the internal looper, which will be on a different thread
from where the test calls close(). If the callback arrives shortly
before close() is called, the callback could be accessing the codec
after it's released.

Synchronize the callback and close() with a lock.

bug: 38213014
Change-Id: Ifa85459203574d257ba27cce471b5b23f0c1c4da

Former-commit-id: 0e8a8a026138d4a2f20daaadd70c9fb0132ab8a3