aaudio: Add more tests for the input stream
Add new tests:
 - AAudioInputStreamTest.testStartReadStop
 - AAudioInputStreamTest.testReadCounterFreezeAfterStop
 - AAudioInputStreamTest.testPauseAndFlushNotSupported

Also found that the implementation of sleep by waiting for the
state change in AAudioOutputStreamTest.testWriting wasn't
working because the initial state was specified wrongly. After
replacing with a real sleep, it has been discovered that it's
impossible to guarantee state change detection when AudioTrack
is used as an underlying implementation of AAudioStream. As a
result, the check for read frames counter to freeze on pause has
been disabled.

Bug: 37755299
Bug: 33354715
Test: passes
Change-Id: I3d3dae0153470e7e6f4c5655df8cc0658f6a9200

Former-commit-id: f93e7a8506941195f4a000091424681d58e6320c