Verifier: remove firstApiLevel check in encryption test
We need all API 25+ (N-MR1 and later) devices to support
ENCRYPTION_STATUS_ACTIVE, regardless of the API level at which the
device was initially shipped.  Therefore we remove the firstApiLevel
check in the test in order to cover devices that were initially shipped
on API 24- but are running a N-MR1 system or above.

Bug: 27699209
Change-Id: I9d5241eef14683ee899bb8eae24342c6bc04f912

Former-commit-id: 8e5063b467e46c11ab0d42f740899724dd66f764