Disable flaky friction test
This CL disables the friction test for AbsListView.

The test sends fling gestures to devices and compares how much
the list scrolls based on the friction.

Unfortunately, the CtsTouchUtil used by the test uses sleeps
between motion events so it is not guarateed to send the same
events between two different iterations, which makes the test
super flaky on the cloud devices.

To unblock oc-dev, I'm disabling the test and we should rewrite
the test not to rely on timers.

Bug: 62133333
Test: CtsWidgetTestCases
Change-Id: I6c136f352a7a179d7d364b847ecb3d16323f774c

Former-commit-id: 5a8de0990a568e480a0f2173a81ff0f1d098b211