Batterystats CTS test waits for app to enter bg.
Some Batterystats CTS tests require the test app to be in the
background (or foreground) state. If the app is in the wrong state, the
test cannot pass, so the app just quits. However, it seems that certain
devices take longer to enter the correct state, causing the tests to
usually fail on those devices. This cl fixes that: if the app
is in the wrong state, it waits for one second and tries again, repeating
several times before finally giving up. As long as the app enters the
correct state within this wait time, the test can proceed as
planned.

Fixes: 38397188
Test: run cts-dev -m CtsIncidentHostTestCases -t com.android.server.cts.BatteryStatsValidationTest
Change-Id: I6dd7048bf3f93d0f31ec90f37f611f695299b34e

Former-commit-id: 29e08dca71bebf0ac8b7ab57f9d780621cb13849