Avoid duplicate test names in CtsLibcoreJavaUtilCollectionsTestCases
Provide a unique name for the test suite created to test a
LinkedList as a Queue to differentiate it from the suite created
to test it as a List to avoid duplicate test names being
created.

Ran the following before and after the fix and checked that
there were duplicates before fixing and none after.
am instrument -w -r --abi arm64-v8a -e debug false \
    -e log true -e timeout_msec 1200000 \
    -e class libcore.java.util.suite.OtherCollectionsSuite \
    libcore.java.util.collectiontests/android.support.test.runner.AndroidJUnitRunner

Bug: 38413440
Test: as above
Change-Id: I5d96ad25edf8a454301004437a65c389c1858da3

Former-commit-id: 2c1cc2432122707e27dd3e485daa48e4f37e4547