Add CTS tests for always-on VPN opt-out.
We add two host-side CTS tests for the new feature "opt-out for
always-on VPN". The tests verify that
DevicePolicyManager#setAlwaysOnVpnPackage() throws an exception when an
unsupported VPN app is about to be set as always-on, and works correctly
for supported VPN apps. It also checks
DevicePolicyManager#getAlwaysOnVpnPackage() behaves correctly under
these circumstances.

Bug: 62503664
Test: run cts-dev --module CtsDevicePolicyManagerTestCases -t com.android.cts.devicepolicy.MixedManagedProfileOwnerTest#<test>
      where <test> is
      testAlwaysOnVpn,
      testAlwaysOnVpnLockDown,
      testAlwaysOnVpnPackageUninstalled,
      testAlwaysOnVpnUnsupportedPackage,
      testAlwaysOnVpnUnsupportedPackageReplaced
      and repeat for MixedDeviceOwnerTest, MixedProfileOwnerTest

Change-Id: Ib381800c5a33221c05d8aafe59eb93bae9788572

Former-commit-id: 5c11c626d74badd05c1864d009c5adc6911b1e7d