Partially fix CTS Preference tests for wearable devices.
This CL tackles one of the many problems there are with this test on
wearable devices. In particular it ensures that while searching for
a particular view the container view is scrolled before it is concluded
that the view does not exist. In this test, the size of the screen was
not taken into account so on small screens of wearable devices it would
not find elements of the list that are not yet loaded (as they are off
screen).

It also ensures that, beforing taking a screenshot, it will
wait until the UI is stable enough to avoid glitch.

Meanwhile, we crops 20% of the screenshots for watches from the right
hand side. This is because the scroll bar on watch is on a curve and
occupy more spaces.

Bug: 37646009
Test: Run CTS tests on bullhead and swordfish.
Change-Id: Icd7822c22b2f5d313b87ef7849cddabd6488d9bf

Former-commit-id: 8b5da0f7e616a46ffae5b2751db2ca8b69db2be1