Add CTS tradefed target preparer for backup tests.
BackupPreparer is used to enable backup and select LocalTransport
on the testdevice before all the tests and return the device
to the original state after all tests are finished.

This is better than the current way of implementing it in the
base test class, which is enabling/disabling backup before/after
each test, which can trigger the device init operation
that can disrupt the execution of the next test since it's run
async and there is no guarantee on when it's going to be executed.

Bug: 62253989
Test: make cts -j32 && cts-tradefed run cts-dev -m
CtsBackupHostTestCases

Change-Id: Iaeb9c62dc77ffce4d96bf992692c7ec9e9caffb3

Former-commit-id: 8b71b0ca70f4c17048c7fcebd13588ffd8dcee99