Set parole_duration=0 for CtsHostsideNetworkTests.
Some of the tests in CtsHostsideNetworkTests module will put the
test app in App standby mode and check the network availability.
If the device enters into parole state during the test, then the test
fails. Set the parole_duration to 0 so that it won't parole state
won't affect these tests.

Bug: 63140218
Test: cts-tradefed run singleCommand cts-dev -m CtsHostsideNetworkTests -t \
      com.android.cts.net.HostsideRestrictBackgroundNetworkTests

Change-Id: I8158eba74f03e82566de1de9f94b981b17175df5

Former-commit-id: b899444714f84aadea446bf76a3ff254481249d4