Fix CTS testAutoSizeCallers_setText for android wear
Reduce the layouts to 200x200px in order to easily
accomodate the most low resolution watches (in compilance
with https://material.io/devices/).

Relax the assertions: we add/remove text 10 times but
on certain devices auto-size will not necessarily produce
10 distinct sizes and this is valid. Be strict with the
last assertion only when comparing to the initial size.

The test is inlined with the support library version now.

Bug: 62878673
Bug: 37645436
Test: cts-tradefed run cts-dev -m CtsWidgetTestCases -t
      android.widget.cts.TextViewTest

Change-Id: I793c6b0fa3946896e457f5238d78b4507c222781

Former-commit-id: 69cb772e662c995d30c7c2046a7be06692bd096b