Fix Fragment.onCreate() not being called in some cases.
Bug 62798751

When an application process has been killed with a retained instance
fragment on its back stack, and a configuration change causes
state save, the retained instance fragment was not calling onCreate()
when it was popped back into existence.

The cause of the problem is that the mRetaining flag was being set,
and when it was supposed to go through onCreate(), it was skipped.
This CL adds tests.

Test: ran fragment tests
Support Lib Change: Ia66a1f474d815f657b6962097139c796f7fa198b
Framework Change: Ia7878f4ac67ee8cfa05903ba6749ab9851f603ee

Change-Id: I512f8c174f557ee7ed18a9b9dc5c0c3e7a5843a5

Former-commit-id: f49e45157f0f8fea93343e199dd83e928260fb3d