Add test for always-on VPN working across reboot.
In N, we added the always-on VPN feature which allows VPN apps'
connection to be persisted across system reboots. We now add CTS tests
to verify this behavior.

More specifically, we set a VPN app as always-on, then reboot the
device, and finally verify that the VPN app is started and that network
connection is tunneled through the VPN after the credential-encrypted
storage of the device / work profile is unlocked.

Bug: 27714439
Test: run cts-dev --module CtsDevicePolicyManagerTestCases -t com.android.cts.devicepolicy.MixedDeviceOwnerTest#testAlwaysOnVpnAcrossReboot
Test: run cts-dev --module CtsDevicePolicyManagerTestCases -t com.android.cts.devicepolicy.MixedProfileOwnerTest#testAlwaysOnVpnAcrossReboot
Test: run cts-dev --module CtsDevicePolicyManagerTestCases -t com.android.cts.devicepolicy.MixedManagedProfileOwnerTest#testAlwaysOnVpnAcrossReboot
Change-Id: I277d621030a49e78c89077d29c320fe90b52aed7

Former-commit-id: 2bba0378cf7818c296fcaaf64951f047f61a6946