Added test to ensure screen is not turned back on with a relayout.
The test case added will turn the screen on in onStart. The onStop
method has a call to trigger a relayout. This test ensures that
the relayout triggered in onStop doesn't cause the screen to
get turned back on again. This is only an issue on phones that
have always on ambient display set.

Fixes: 64139966
Test: cts/hostsidetests/services/activityandwindowmanager/util/run-test
CtsServicesHostTestCases
android.server.cts.ActivityManagerActivityVisibilityTests#testTurnScreenOnActivity_withRelayout

Change-Id: I7d9932a98a9dc81777a12154e273057e0b32f2fd

Former-commit-id: c3922dca5119b86ea0bfd19bf819cf29bfe2a532