Use all certs for computing package signing sha256 - CTS
In several places we compute the sha256 of the app's signing
certificate (instant cookie storage, backup account permission
grants, static shared lib matching). It is possible that an app
is singed with multiple certs which unfortunately can appear
in a random order. We were using only the first certificate to
compute the hash which may be problematic for apps signed with
multiple certs which are later reordered. If an app update's
certs are reordered for cookie storage the app would not be able
to access the cookie, for account grants the app would not get the
grant, and for shared libs the app would fail to install due to
a missing lib.

Test: all cookie CTS tests pass
      all static shared lib CTS tests pass
      added test that cookie data not lost on sha256 computation change
      added test that lib instal works when specifying multiple certs

bug:64270295

Change-Id: I39919dbaf246e0ecc36a1e953faa0910aa377bfe

Former-commit-id: 7107c9bb615c892d5662d5487c3fe6df9236720a