WebView: change getSafeBrowsingEnabled test
getSafeBrowsingEnabled()'s return value depends on if the user has opted
into Safe Browsing (as determined by
SafetyNetClient#isVerifyAppsEnabled()). Since we can't control GMS or
the Google setting in this test, we should make no assumptions about the
setting's value.

Given this, the only thing we can safely assert is that Safe Browsing
will be disabled by default, and can be explicitly disabled via the API.

Bug: 64095293
Test: run cts -m CtsWebkitTestCases -t android.webkit.cts.WebSettingsTest#testEnableSafeBrowsing
Change-Id: I6bd2fb26c889f17f8f98d83b9c36c9bd3d03f728

Former-commit-id: 48a7ca45537b39ecd9fc2717f843b361ff11d712