assert mediaextractor uses a seccomp filter
As part of the media frameworks hardening effort, mediaextractor
runs in an unprivileged sandboxes intended for hosting risky
code responsible for parsing file formats. By default, the extractor
process runs with a seccomp filter which limits the syscalls
available to the process, providing additional protection for the
kernel and other processes. Enforce that these processes exist and
use a seccomp filter.

Test: cts-tradefed run cts -m CtsSecurityHostTestCases -t \
    android.security.cts.ProcessMustUseSeccompTest
    android.security.cts.ProcessMustUseSeccompTest#testMediaextractorHasSeccompFilter
    pass
Bug: 64536899
Change-Id: Ibfcf0942cb611e1ab26463293eed966e4ed91e81

Former-commit-id: e6dfef9b9248a6f87d939ada97cb95431c679255