CtsMediaBitstreamsTestCases: bitstreams package via annotation
There are two configurations for the CtsMediaBitstreamsTestCases.
The default configuration is "standard", which contains 10-12 bit
bitstreams in codecs h264, hevc, vp8, and vp9. The "full"
configuration is a superset of "standard", and about 50 times larger.
The default configuration is "standard".  Use the annotation
android.media.cts.bitstreams.FullPackage to select the full subset.

Bug: 64342361
Test:  cts-tradefed run cts -m CtsMediaBitstreamsTestCases --module-arg \
  CtsMediaBitstreamsTestCases:include-annotation:android.media.cts.bitstreams.FullPackage

Change-Id: I1e78f556b0e136c63faa86c83b647e05b7018f79

Former-commit-id: e847ee15fbceb64a159f25a005c8f21aee6d062b