[DO NOT MERGE] Wait for all activities to be stopped.
There's a chance that even though the power dumpsys
shows that the screen is off, the activities might still
be started. This is an additional check to ensure that
all activities are stopped before launching the activity.
The check that display is off remains in the code so this
should guarantee that both the screen is off and no activities
are started.

Test: ActivityManagerActivityVisibilityTests#testTurnScreenOnAttrRemove
over 50 times without failing.
Fixes: 65646519

Change-Id: Idcdceb4c98c0688c58cbead572567d5f50999756

Former-commit-id: 745a60951d2c47325b0328a6bf93a1f3760d02d4