Increase timeout on decoding accuracy tests
increases the timeout for the accuracy tests to ensure that some of the
older (slower) devices have time to process huge (3840x2160) frames.
The tests decode only 10 frames, so startup time is a more significant
part of the test and had not been well accounted for in the calculation
of what the timeout should be.

Bug: 66144372
Test: ran affected CTS tests on bullhead
Change-Id: I3ca03f85dab4b4aa89f349d726e69aa487151824

Former-commit-id: 19cf86f1a04f7326b3c8e68591c325557e7d642d