Start the installer activity in new task
The dialog is actually there. You can observe the dialog if you wait
for a while (few seconds).

The activity is actually started (and resumed) right away, but for
some reasons, it is not rendered until a few seconds are passed.
Feel like it may be a bug in in the rendering side...

Work around this by starting the activity in a new task, this is not
something this test is testing for anyway.

Fixes: 63814740
Test: Run this test in CtsVerifier.

Change-Id: I9d1e5c70773c5cc6811d26811d916236dcd10179

Former-commit-id: 16a01c0a4ea01c3ca1447baec301779b83d91791