CtsLiblogTestCases liblog#wrap_mode_blocks failing (DO NOT MERGE)
Angler and Bullhead userdebug builds suffer from less than 30 seconds
logspan in the main log buffer.  The result is all bugreports (which
typically start their logcat dumps about a minute into the request),
and any tests that depend on the logs not being DOS'd are useless or
dysfunctional.

In this case liblog#wrap_mode_blocks fails because the main log buffer
is under attack and thus returns prematurely. This likely means there
is an increase in battery draw too.

The fix is to tame the platforms and the potpouri of Cts tests that
also contribute to the logging pressure.  It should be noted this test
passes on user builds.

Test: none
Bug: 67377433
Change-Id: Idaeddcfdf5791c876ad42504762e4cf31492b38e

Former-commit-id: 0dfd7109b574ee2a19a662db7615d9798a7971ef