add more execution time to embedded devices in media projection tests
This CL makes the media projection tests in CTS allow 1 fps capturing.
Embedded devices have no GPU support and low performance CPU on the
board. Media projection works include media encoding and capturing a
screen at the same time. Those works depend on only CPU because of the
non GPU support on the board. So, the performance expectation of those
devices is at the above condition.

Bug: 65864256
Test: android.view.cts.SurfaceViewSyncTest on imx7d and walleye
Change-Id: Id0b8f2e62b26e08f2515e79d5161118808cb70c2

Former-commit-id: 190cd06659e0fc4595596b74eebfd03df1d16bca