Modified test to assert switch to new non-guest user and back
Modified the test to assert no errors when both going to new user and
going back to the original user. The user switch to any running users
now checks for the log string for "Continue user switch" instead of
"Boot completed" which continues to check for switching to a new user.
Continue switch user would be the final message to indicate that the
user has successfully switched and onUserChanged callbacks are fired
(for switching to existing user). Also check the logs even after the
exit string is found since user change callbacks could crash. Finally
added a test for new non-guest users.

Test: atest CreateUsersNoAppCrashesTest
Bug: 68929438
Change-Id: I82b735a3eb11b11df0e1022f717381d688418c36

Former-commit-id: 7e54cd2a1b5f0a7b8f88f1a58c0127fd5fcfb005