Regenerate ThemeHostTest reference images with local system image
Implement a "local" option in generate_images.py, which runs emulator
without a SDK and it is using a locally build system image.
Generate new reference images for ThemeHostTest to resolve differences
in antialiased pixels. Images were generated using a system image from
a local build synced with a manifest file of ab/4474236.

Bug: 67866323
Test: passed ThemeHostTest.
Change-Id: I74d478379573461310a9f3567ec87add089eb079

Former-commit-id: 3baab43a763a3f78568604aade31b59c0bdab49e