CTS: Use activity context when obtaining resource dimension info.
Using the application context to request resource dimension info will use
the primary display configuration. This can break for activities running
on a virtual display with different density settings.
Instead, the test should use the activity context to assure it requests the
resource dimensions based off the display the activity is running on.

Test: Validated CtsWidgetTestCases pass
Bug: 69370835
Change-Id: I56f844f77907750d492d5be588b168f26d26640b

Former-commit-id: 2a3794e2064a9d22daf2c435060138bfbd12d82f