Allow Mock IME to disallow fullscreen mode
By default InputMethodServie (IMS) triggers fullscreen mode when the
device orientation is landscape.  For end-to-end tests, however, we
should be able to explicitly control the condition when the Mock IME
starts fullscreen mode because for some form-factors such as tablets
and large screen devices.

With this CL, fullscreen mode is disallowed by default in Mock IME.
To explicitly allow Mock IME to enter fullscreen mode, specify
option when creating MockImeSession as follows.

  try(MockImeSession imeSession = MockImeSession.create(
        InstrumentationRegistry.getContext(),
        InstrumentationRegistry.getInstrumentation().getUiAutomation(),
        new ImeSettings.Builder().setFullscreenModeAllowed(true)) {

      // Use imeSession to interact with the Mock IME here.

  }

Bug: 69845539
Test: atest CtsInputMethodTestCases
Change-Id: I3a108a9e776f62496ac0ab809536b1cec197e60f

Former-commit-id: cfd2689ca09f0ea9d58b67434124604d83c4e3f6