Explicitly pass Locale.ENGLISH to SimpleDateFormat constructor.
If not provided with a locale, SimpleDateFormat (like the rest of
java.text.*) will use the default FORMAT locale, which can be anything.

All other output from ResultHandler is in English so the date included
in the output should be in English as well (and not in some unknown
locale deduced from user settings, that might or might not be English).

Test: cts-tradefed
Change-Id: If5925589d0b4b5ed24e6b946502e90c9858e35fa

Former-commit-id: 8df1e4aa3452b7be295a816c9e26a5f1715f8209