DO NOT MERGE Fix for testConfigurationUpdatesWhenRotatingToSideFromDocked
When the divider position is adjusted to a non-square frame,
it might be resized half and half after rotate the device depends on snap-algorithm which determines the divider position.
If initial app's rect is almost square (aspect ratio is less than 1.1),
adjust the divider position on each rotation to make sure activity gets non-square frame.

BUG=69867833

Change-Id: I26eecf9e5e44a4ed3938e199ce3f22c564dc84fa

Former-commit-id: c5d98c80546e9ac906c1facd38bb3d74025cf777