Fixing failing KeyStoreTest CTS test
An underlying implementation change in BCRSAPrivateCrtKey caused it to
default reject equality with any non RSAPrivateCrtKey's in its .equals()
method.

This change will check for instances of RSAPrivateCrtKey's in either the
expected or actual key and instead compare them based on equality of
modulus and private exponent, exactly like OpenSSLRSAPrivateKey handles
equality checking for objects that are of a different RSAPrivateKey
type.

Bug: 68057941
Test: KeyStoreTest CTS tests all pass
Change-Id: I7e47e9c95b748b6e2627f3bab740bf3fd45c966d

Former-commit-id: ff83f942cf63c5220f16347ad9538f9d1b406a7d