Add tests for WebView.disableWebView().
Add tests for the ability to disable WebView.

This introduces a new test framework for WebView tests across multiple
processes. Several WebView tests need to run code in a freshly created
process, or in more than one process simultaneously. This is achieved by
defining two services which each set android:process to a different
value in the manifest. The services will run any method they are told to
run by the client, and will exit when told to. Exceptions thrown by the
method calls will be propagated back to the client, allowing the use of
normal junit assertions.

Bug: 63748219
Test: cts android.webkit.cts.WebViewDataDirTest
Change-Id: Iddc3faea96c5af799bfd5057c8bb01334e613722

Former-commit-id: fc7adf379f2bf33c413b5b9f8f8b423c04064ec3