MotionEventTest: Expect no events in VR mode
The MotionEventTest creates an application overlay window and
checks the events it received.  In VR mode such windows are not
allowed, which is allowed by the CDD.  By not allowing these
events, the intent of the test is satisfied (no information is
leaked), so this CL adds a VR case that anticipates this
behavior.

Bug: 64533803
Test: Run android.security.cts.MotionEventTest#testActionOutsideDoesNotContainedObscuredInformation
Change-Id: Ibde12159bffaaa3f474ab1bcfae19168a5651624

Former-commit-id: e245853f3febbf336cf4016e151bfbe7c05731f7