Fix minor WaitForValidActicityState issue
This is a followup CL of I90306d4eba and I8cf271a434.

When ComponentName is used to create WaitForValidActicityState object,
activityName was set to simple class name WITH a leading '.'. This CL will
fix to set simple class name WITHOUT a leanding '.' for compatibility.

There is no functional change because we simple ignore activityName when
componentName is set.
@see ActivityAndWindowManagerState#shouldWaitForActivities(String,WaitForValidActicityState...)

Bug: 70190274
Test: No new test starts failing.
  atest CtsActivityManagerDeviceTestCases
  atest CtsWindowManagerDeviceTestCases
Change-Id: I85369565f2b59224d8ed75f134715144ee50a93d

Former-commit-id: 378c41fc00866b84fc0529e421fdf8b3d5ba7f6f