Add CtsUnofficialApisUsageTestCases
CtsUnofficialApisUsageTestCases ensures that java modules (apks and jars)
installed in the vendor partition are using public and system APIs only.

It downloads the java modules from the device and extracts symbol
references from there. Then the references are compared against the
public and system APIs. If a symbol reference which isn't in the
approved APIs, then it is a test failure.

Bug: 71521429
Test: cts-tradefed run cts --module CtsUnofficialApisUsageTestCases

Change-Id: I7dad838f52fae97db002959af0ef4d744b5c575b

Former-commit-id: c7c9ddc14dd29f237121dea81b767947572da87d