Update CtsAppSecurityTest to use fully qualified names
Most of the tests use fully qualified class names so
sticking to that convention.
See TF related change to re-enforce the actual check of
instrumentation results which caused some tests to show
as PASSED even if it failed.

Test: ./cts-tradefed run cts-dev -m CtsAppSecurityHostTestCases
some failures that are on class untouched by the original issues
(EphemeralTest mostly)
Bug: 68202822

Change-Id: If1e923180eecee16738450f70f27b0c88b1fe0d7

Former-commit-id: 1e821839083637c6ca4380a94b5a2424b594cee8