Remove overlapping static and dynamic libraries
Adding libraries to both LOCAL_JAVA_LIBRARIES and
LOCAL_STATIC_JAVA_LIBRARIES creates duplicate classes which causes
problems for Proguard.

The bouncycastle library is removed from LOCAL_STATIC_JAVA_LIBRARIES
because it is available at runtime on the bootclasspath.

The cts-security-test-support-library is removed from
LOCAL_JAVA_LIBRARIES because it is not available at runtime.

Bug: 30188076
Test: make CtsVerifier
Change-Id: If0d8a3129bb02e42c48abaf87dcf80a456b0d788

Former-commit-id: 3e1528702b923cac2e9845ffbb9575f1dadbccda