Fix broken CTS test: WebSettingsTest#testAllowMixedMode
Due to a change in chromium [1] the "localhost" addresses
cannot be treated as mixed content anymore and since the
test webservers use "localhost" the test is broken.

This patch fixes the issue by using a combination of a ssl
webserver and a custom shouldInterceptRequest implementation
to intercept requests to insecure url addresses. In particular
the ssl webserver attempts to load insecure resources
(javascript file or an image) and depending on the mixed content
settings in webview these requests are either allowed or not.
This is reflected in the calls to shouldInterceptRequest.

[1] https://chromium-review.googlesource.com/702277

BUG: 71992132,68877177
Test: cts-tradefed run cts -m CtsWebkitTestCases --test android.webkit.cts.WebSettingsTest#testAllowMixedMode

Change-Id: I28119fd0f7a7e9c97212f3e1d9c2c6cb57ac2fa6

Former-commit-id: 89f0fc3d176b59fa52efebc304a515d58935e866