Update testCreateTransform() for one-way transforms
This patch augments testCreateTransform() to
selectively apply the ipSecTransform in one/both
directions and then verifies that the link succeeds or
fails correctly. In addition, this converts the test
to use the Java sockets API for DatagramSocket instead
of using the FD-based API. One of the combos does not
fail as expected, so b/72048263 is tracking further
work needed.

Bug: 72047396
Bug: 71717213
Test: cts - IpSecManagerTest

Change-Id: Ied961f442b29f1e66ded37c7ffe7afdd4a5c49da

Former-commit-id: 78e1d4575d4f3aa2ccf865d218074a41719eda48