cts: Wait for VFS cache to expire
In the test suite testMultiViewMoveConsistency(), the function testMoveBack()
moves a file from an app with RW permissions to an app with RO permission.
The test tries to "stat" the new file immediately after moving it.
The VFS cache has not expired yet leading to older UID returned for the test
on ARM_64 Kernel 3.10.

Similar example:
Test testMovePackageSpecificPaths() is similar to this test but we can see that
there is a sleep added in testMovePackageSpecificPaths() after Os.Rename()
for VFS cache to clear.

We need to wait for VFS cache to clear before calling "stat" for the new file
in the test testMoveBack()

Change-Id: Ic810cd6f47d4bfc73607393f44558cce4c3e78ca

Former-commit-id: 2cc3ae4127c08ab8892d161fce4c0cc6ee0db6d2