Allow Keystore tests to interact with lock screen
Certain keystore features rely on behavior changes when the screen is
locked or unlocked, but so far there has been no way to lock or unlock
the screen, or set or unset passwords, from a CTS test. This change
pulls in the classes from activity manager tests that interact with the
screen and various activities, giving an easy interface for locking the
screen.

Test: Added testCanCreateAuthBoundKeyWhenScreenLocked, which locks the
screen and confirms that it is possible to generate auth-bound keys, and
testCannotUseAuthBoundKeyWhenDeviceUnlocked, which tries to create auth
bound keys on an insecure device and makes sure it throws the correct
exception.

Bug: 67752512

Merged-In: I388633ae5d3e8ce60c35dfef2eeeb84ecd143bb1
Change-Id: I388633ae5d3e8ce60c35dfef2eeeb84ecd143bb1
(cherry picked from commit 930d1c4432be28935d2ee3952b09d24ed46dc3c0)

Former-commit-id: 2dfb8a4dd69a2139f85b9fe769255507390cf0ec