Fix battery stats validation CTS tests
The tests assume that the newly-installed app will be able to
run jobs and syncs immediately.  However, newly-installed apps are
now treated as being in the bottom standby state until the user
interacts with them directly.  Fix the test by explicitly hoisting
the app into the fully active standby state as part of the install
sequence.

Fixes: 72063927
Test: atest com.android.server.cts.BatteryStatsValidationTest
Change-Id: I1bbb57573ed804a4b753426a5213ebef70bfd3c3

Former-commit-id: c35b85701d920965a1321a851682f547245fe94b