appsecurity: Add tests proof-of-rotation sharedUid key rotation.
Shared UID apps have special considerations when rotating signing
certificates since they may run in the same security domain.
Rotation is allowed provided that the signing certificates remain
the same or that the newer one grants the sharedUid capability to
the old signing certificate(s).  Make sure installs are sucessful
when granted and not otherwise.

Bug: 64686581
Test: Builds, tests pass.
Change-Id: I206b2ad3f7985c53fc672249f92bc330388f8f73

Former-commit-id: b47ab5e8af31228afdbcf44f1a65edf0af26abbc